+++ apps/wallop_core/lib/wallop_core/resources/draw.ex
@@ -35,6 +35,7 @@ defmodule WallopCore.Resources.Draw do
 
       change(set_attribute(:api_key_id, actor(:id)))
       change(set_attribute(:status, :open))
+      change({WallopCore.Resources.Draw.Changes.IncrementApiKeyDrawCount, []})
       change({WallopCore.Resources.Draw.Changes.ValidateCallbackUrl, []})
       change({WallopCore.Resources.Draw.Changes.RecordStageTimestamp, key: "opened_at"})
     end